requirements.txt


# YOLOv8
ultralytics==8.2.50
opencv-python==4.10.0.84
torch>=2.1.0
torchvision>=0.16.0

# OCR
paddleocr==2.9.1
paddlepaddle==2.6.1

# FastAPI backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Frontend (FastHTML needs Jinja2 + Starlette)
jinja2==3.1.4
starlette==0.37.2

# LLM Support (Hugging Face + OpenAI, optional)
transformers==4.43.3
sentencepiece==0.2.0
openai==1.40.3

# Image Enhancement (GAN/Diffusion)
Pillow==10.3.0
scipy==1.13.1
tqdm==4.66.4
diffusers==0.30.2
accelerate==0.34.2

# Utils
numpy==1.26.4
matplotlib==3.9.2
